+++ src/main/java/racingcar/model/RacingGame.java
@@ -1,9 +1,11 @@
-package racingcar.service;
+package racingcar.model;
 
-import racingcar.domain.Car;
 import racingcar.utils.Randoms;
 
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.Comparator;
+import java.util.List;
 
 import static java.util.stream.Collectors.toList;
