+++ src/test/java/racingcar/CarTest.java
@@ -2,7 +2,7 @@ package racingcar;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.domain.Car;
+import racingcar.model.Car;
 import racingcar.exception.InvalidCarNameException;
 
 import static org.assertj.core.api.Assertions.assertThat;
