+++ src/test/java/racingcar/RacingGameTest.java
@@ -3,8 +3,8 @@ package racingcar;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.domain.Car;
-import racingcar.service.RacingGame;
+import racingcar.model.Car;
+import racingcar.model.RacingGame;
 
 import java.util.ArrayList;
 import java.util.List;
