+++ src/test/java/racingcar/ValidationUtilTest.java
@@ -2,7 +2,7 @@ package racingcar;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.domain.Car;
+import racingcar.model.Car;
 import racingcar.exception.DuplicateCarNameException;
 import racingcar.exception.InvalidCountRangeException;
 import racingcar.exception.SingleCarException;
