+++ src/test/java/racingcar/view/OutputViewTest.java
@@ -2,7 +2,7 @@ package racingcar.view;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.domain.Car;
+import racingcar.model.Car;
 
 import static org.assertj.core.api.Assertions.assertThat;
